Moving files to correct folders
